refactor: use checkModel instead of checkApi for better semantics
